Fix libproxy's shared lib version, ok aja@
cmake.port.mk sets LIBfoo_VERSION for cmake to use as appropriate,
but this only works when the library name matches between SHARED_LIBS
and the cmake infrastructure - in this case there's a mismatch with
"proxy" vs "libproxy".

This used to be handled in this port by patching cmake files but
that's awkward to maintain so no surprise it accidentally got lost
in the last update; replace it by setting appropriate CONFIGURE_ENV
and MAKE_ENV in the port Makefile instead which is more maintainable.

Build problem with dependent ports reported by nigel@.