Update to drush-8.0.3.
This is packaged from a locally generated tarball, built from the downloaded
github archive, plus required dependencies installed via composer. Specifically
not using upstream's normal .phar because we need to patch it so backups work
(it uses options not supported by base tar).

Drop the forced dependency on Drupal 6 as this is suitable for D6/7/8 and is
useful for people who have installed Drupal outside of ports.