This patch introduces pledge(2) to mcabber with the following promises:
rpath/wpath/cpath/fattr: it needs to read, write and create files and set the
correct attributes
inet/dns: network application needs both promises
tty: ncurses is used for the interface so it needs this
proc/exec: this is only needed if events are configured in the config files

In order to apply this we have disabled modules support since it would need to
load libraries dinamically and therefore prot_exec promise would be needed but
that is not a good strategy. This is mentioned in DESCR to reflect the change.

OK tim@ after his suggestions and also OK by the maintainer Markus Hennecke
since no users objected this change.