Import ports/databases/ksql. ok sthen@
ksql is a "lazy man's" wrapper for the SQLite C-language interface. It
is written to make sure sloppy programming won't cause your SQLite
database to be inconsistent (transactions and statements open on exit
and certain signals). Specifically, it exits on any database errors; and
upon exit (or signal), cleans up open statements, transactions, and
databases. See ksql(3) and friends for details.