Import progit-1084.
Pro Git (Second Edition) is your fully-updated guide to Git and its usage in
the modern world. Git has come a long way since it was first developed by
Linus Torvalds for Linux kernel development. It has taken the open source
world by storm since its inception in 2005, and this book teaches you how to
use it like a pro.

Effective and well-implemented version control is a necessity for successful
web projects, whether large or small. With this book you'll learn how to
master the world of distributed version workflow, use the distributed
features of Git to the full, and extend Git to meet your every need.

ok benoit@ krw@