Merge commit 5561199c871293cca0b3dc2c6feff0d7783111c6 from upstream.
Allow alt+tabbed windows over fullscreen

Fullscreen windows should only be on top when they are in focus. Change
the stacking level temporarily back to WMNormalLevel if the fullscreen
window loses focus due to an alt+tab operation.

Change the stacking level back to WMFullscreenLevel if the fullscreen
window receives the focus again.