Move chromium to the new build system (gn - generate ninja)
This commit moves the chromium port to use gn to generate ninja files
which will hopefully resolve the random build errors. Gyp is dead so
this move has to be done anyway.

Also change the way how we handle openbsd. Let's assume we are running
on linux and only handle !linux cases where needed. This also ensures
that a change on linux will for sure end up in our builds as well.