Import facette 0.3.1pre0.
Facette is a web application to display time series data from various
sources — such as collectd, Graphite or InfluxDB — on graphs, designed
to be easy to setup and to use.

Uses npm install during build to fetch dependencies, so bundle them in a
separate tarball and selfhost them.

www: http://facette.io

with feedback & ok ajacoutot@