Import graphics/grafx2
GrafX2 is a bitmap paint program inspired by the Amiga programs Deluxe
Paint and Brilliance. Specialized in 256-color drawing, it includes a
very large number of tools and effects that make it particularly
suitable for pixel art, game graphics, and generally any detailed
graphics painted with a mouse.

OK giovanni@