Repair samba-tool domain provision --use-ntvfs
The situation is a mess.  Upstream says that s3fs (the original smb code
from samba3) requires filesystem ACLs, which we don't have.  The ntvfs
code (new in samba4, but now deprecated) fit the job, but
adding --with-ntvfs-fileserver doesn't actually provide a working 'smb'
service (see "server services" in smb.conf(5)).

So right now it seems that the workaround is to provision
using --use-ntvfs, but then to strip 'smb' from the 'server services'
line.

Reports welcome...