import pcre2-10.22
PCRE2 is a revised API for the PCRE library, which is a set of functions,
written in C, that implement regular expression pattern matching using the same
syntax and semantics as Perl, with just a few differences. Some features that
appeared in Python and the original PCRE before they appeared in Perl are also
available using the Python syntax. There is also some support for one or two
.NET and Oniguruma syntax items, and there are options for requesting some minor
changes that give better ECMAScript (aka JavaScript) compatibility.

initial port by czarkoff@
ok aja@