Import opendkim, a library and milter for DKIM
Port from Johan Huldtgren, ok sthen@

DESCR:

OpenDKIM is an implementation of DKIM (RFC 6376, Domain Keys Identified Mail),
an email sender authentication system. It consists of a library that implements
the DKIM service and a milter-based filter application that can plug in to any
milter-aware MTA to provide that service to sufficiently recent sendmail MTAs
and other MTAs that support the milter protocol.