Update to qgis 2.18.2.
Remove patch-src_gui_qgisgui_h which was part of
https://github.com/qgis/QGIS/commit/3e738405e2e91eaf8a54f82640e5f59fc977e21a
upstream.

Switch to use gcc4 on all platforms (not only on i386), the new DWG
importer (added in a stable branch.. sigh) doesn't build with base gcc,
and what's the point of trying to build a modern c++ codebase with a 10
year-old compiler ?

Bump all shlib majors for lots of symbols removed as a consequence.