import p5-Statistics-CaseResampling 0.15 OK benoit@
Comment:
efficient resampling and calculation of medians

Description:
The purpose of this (XS) module is to calculate the median (or in
principle also other statistics) with confidence intervals on a
sample.  To do that, it uses a technique called bootstrapping.  In
a nutshell, it resamples the sample a lot of times and for each
resample, it calculates the median.  From the distribution of
medians, it then calculates the confidence limits.