Add a port for the closed firmware required to boot Raspberry Pi systems.
It is loaded from a FAT16 or FAT32 filesystem on removeable media,
the hardware does not accommodate loading it from a ROM.

This includes software that runs on the VideoCore part of the Broadcom
SoCs which boots the arm cores and continues to provide services to
the arm cores after they have booted.

Redistribution is permitted subject to the conditions outlined in
https://github.com/raspberrypi/firmware/blob/master/boot/LICENCE.broadcom

feedback from and ok sthen@ ok phessler@ on an earlier version.