import py-wheel-0.30.0a0 needed for a incoming update for py-argcomplete from jsg@/sthen@
ok/tweaks shadchin@

A wheel is a ZIP-format archive with a specially formatted filename and
the .whl extension. It is designed to contain all the files for a PEP
376 compatible install in a way that is very close to the on-disk
format. Many packages will be properly installed with only the “Unpack”
step (simply extracting the file onto sys.path), and the unpacked
archive preserves enough information to “Spread” (copy data and scripts
to their final locations) at any later time.