Fix build breakage by shoplifting a patch from tor-browser (see https://trac.torproject.org/projects/tor/ticket/21514) - patch pointed at by Jan De Mooij@mozilla.
In bug #1334933, as part of a security fix the code for W^X not
merged/rearranged, but was removed - thus triggering mmap protection
during build on OpenBSD, as reported by naddy@. My bad for not spotting
this.

The tor-browser patch just reinstates some W^X bits, instead of marking
TB as needing wxallowed. This is only temporary for 45, 52 won't need
that anyway.

Tested building and working fine.