import productivity/vdirsyncer, from Remi Locherer, ok danj@
Vdirsyncer synchronizes your calendars and addressbooks between two
storages. The most popular purpose is to synchronize a CalDAV/CardDAV
server with a local folder or file. The local data can then be accessed
via a variety of programs, none of which have to know or worry about
syncing to a server.