Update to sblg-0.4.1.
Changelog:

"Enpower -l (tag-listing mode). This is now able to emit all tags in JSON
formatted output via -j, which makes it easy for dynamic sites to play with
tag listings in any way they choose. (See sblg(1) for an explanation of the
output format.) Also add -r, which stipulates that -l will print tag-first.
This is most useful for -j, as it allows for easy browsing by tag name."