new update-patches.
that one deals with .beforesubst files, so it should be nicer with
SUBST_CMD handling. It also doesn't care nor emits timestamps markers
in diff, so less garbage to think about.

Please report any issue, things work better so far, but it's a work
in progress and change to make it better.

'let's finish this in tree' okay aja@