Import py-simpleeval-0.9.5 from semarie with minor tweaks from me
A quick single file library for easily adding evaluatable expressions into
python projects.

Say you want to allow a user to set an alarm volume, which could depend on the
time of day, alarm level, how many previous alarms had gone off, and if there
is music playing at the time.

Or if you want to allow simple formulae in a web application, but don't want to
give full eval() access, or don't want to run in javascript on the client side.