import p5-List-MoreUtils-XS 0.418 p5-List-MoreUtils has been split in two modules upstream. Better use the software in our ports tree as build dependency, patch away the included modules. OK sthen@
Comment:
provide compiled List::MoreUtils functions

Description:
All of the functions are implementable in only a couple of lines
of Perl code.  Using the functions from this module however should
give slightly better performance as everything is implemented in
C.  The pure-Perl implementation of these functions only serves as
a fallback in case the C portions of this module couldn't be compiled
on this machine.