Update Calligra to 3.0.1
From release notes:

"We have chosen to cut back on the number of applications. Krita has left us to
be independent and although it was emotional it was also done with complete
support from both sides. We are saying goodbye to Author, which never
differentiated itself from Words. We also removed Braindump the purpose of
which will be better fitted by a new application (nothing planned from our
side). Flow and Stage has gone in this release but we intend to bring them back
in the future. And Kexi has own release schedule but is still part of the
Calligra community."
-- https://www.calligra.org/news/calligra-3-0-released/

Many hints and tweaks from lendry@, sthen@ and jca@:

- Remove some kf5 libs from LIB_DEPENDS. They are only used
  as RUN_DEPENDS. spotted by  jca@

- Add devel/gettext as RUN_DEPENDS and libiconv as LIB_DEPENDS.

- Make the update process possible for editors/calligra-l10
-- idea from landry@ Thanks!

... and some more. Thanks!

ok landry@