Use iconv(1) to convert the russian manual page from KOI8-T to UTF-8 before installing; sthen@ agrees with the general direction.
Even though the russian manual is known to be heavily outdated, i
strongly dislike the idea of expecting porters to judge the quality
of the content of third-party documentation (as opposed to fixing
technical issues).  When i said as much on ports@, nobody really
objected.

So for now, merely fix the encoding.  If people *really* want the
manual page deleted, that can still be done.