Remove chinese/crxvt.
We recently removed mrxvt and rxvt due to the fact that upstream is
dead and that thee code is vulnerable to CVE-2017-7483.

On top of that, crxvt is very unstable: it dumps core after entering
a few commands or pressing enter repeatedly a couple of times.

OK jca@, dcoppa@, bcallah@, kevlo@, bentley@