Import obfs4proxy-0.0.7.
obfs4proxy is a tool that attempts to circumvent censorship by
transforming the Tor traffic between the client and the bridge. This way
censors, who usually monitor traffic between the client and the bridge,
will see innocent-looking transformed traffic instead of the actual Tor
traffic.

obfs4 (a.k.a. The Obfuscator) is a transport with the same features as
ScrambleSuit but utilizing Dan Bernstein's elligator2 technique for
public key obfuscation, and the ntor protocol for one-way
authentication. This results in a faster protocol.

From MAINTAINER attila.

ok sthen@