import ruby-gpgme:
Ruby-GPGME is a Ruby language binding of GPGME (GnuPG Made Easy). GnuPG
Made Easy (GPGME) is a library designed to make access to GnuPG easier
for applications. It provides a High-Level Crypto API for encryption,
decryption, signing, signature verification and key management.

it's a requirement for the hiera-eyaml-gpg backend.

Quite some help from and OK jeremy@