Import lang/cparser, a C99 parser and frontend for libFirm. ok bentley@
cparser is a recursive descent C99 parser written in C99. It contains a
preprocessor, lexer, parser, constructs an AST and does semantic
analysis. It acts as a frontend to the libFirm intermediate
representation library. This way optimization and code generation is
performed. The compiler supports cross compilation to multiple target
architectures with a command-line switch. It comes with driver logic for
calling assemblers and linkers as well as parsing command-line options.
This allows it to be a drop-in replacement for gcc or clang in many
situations.