Update lang/swi-prolog to version 7.6.0.
For the first time, all tests are now passing on OpenBSD. This is thanks to the
hard work of Jan Wielemaker who went on a bug fixing rampage after I set him up
an OpenBSD VM. As a result, lots of patches also disappeared, thus making the
port simpler. Many thanks for this Jan!

OK jeremy@, thanks.