Import games/hyperrogue, a roguelike game in a non-Euclidean world. ok schwarze@
You are a lone adventurer in a strange world, where geometry does not
work in the expected way. Gather as much treasure as you can before the
nasty monsters get you. Explore about 50 different worlds, each with its
own unique treasures, enemies, and terrain obstacles. Your quest is to
find the legendary treasure, the Orbs of Yendor. Collect one of them to
win! Or just ignore your quest and collect smaller treasures.

The twist is the unique, unusual geometry of the world: it is one of
just few games which takes place on the hyperbolic plane. Witness a grid
composed of hexagons and heptagons, straight lines which seem to be
parallel, but then they diverge and never cross, triangles whose angles
add up to less than 180 degrees, how extremely unlikely is it to reach
the same place twice, and how the world seems to be rotated when you do
return. All this matters for the gameplay. The game is inspired by the
roguelike genre (although in a very minimalist way), works of M. C.
Escher, and by puzzle games such as Deadly Rooms of Death.