Import freedoom-0.11.3.
Freedoom is a freely licensed game pack for the Doom engine, split in two parts.

Phase 1 is the first part of the single-player game, containing four chapters
which are nine levels each, smoothly paced for beginner players. It is
compatible with mods for the original Doom and The Ultimate Doom.

Phase 2 is the second major part of the single-player game, a massive 32-level
chapter expanding upon the same concepts present in Phase 1, as well as
containing additional monsters and the double-barreled shotgun! The levels in
this game tend to demand more skillful play than the previous chapters. It is
compatible with mods for Doom II.

FreeDM is a fast-paced competitive deathmatch game, part of the Freedoom
project. Rather than the usual single-player focused levels, these contain
no monsters and are intended for deathmatch only. It is compatible with mods
for Doom II.

ok landry@