Import lang/cim, the GNU Simula compiler. ok juanfra@
GNU Cim is a compiler for the programming language Simula (except
unspecified parameters to formal or virtual procedures). It offers a
class concept, separate compilation with full type checking, interface
to external C routines, an application package for process simulation
and a coroutine concept. The compiler and run-time system are written in                                                                        
C, and the compiler produces C code that is passed to a C compiler for
further processing towards machine code.