Import x11/wmutils, a set of tools for X window manipulation. Submission by Kaashif Hymabaccus <kaashif AT kaashif DOT co DOT uk>, who takes MAINTAINER, thanks! ok bentley@
wmutils is a set of tools for X windows manipulation. Each tool only has                                                                        
one purpose, to make it as flexible and reliable as possible. This
package includes tools to list, kill, move, teleport, focus and resize
windows, and more.

wmutils opt contains programs to move windows with the mouse, control
window borders, and view a window's events and title. These programs
rely on wmutils-core to be useful, but can be run without it if you have
another method of getting window IDs to operate on.