Import games/openjk, an open source reimplementation of Jedi Academy game. ok bentley@
OpenJK is an effort by the JACoders group to maintain and improve the
game engines on which the Jedi Academy (JA) and Jedi Outcast (JO) games
run on, while maintaining full backwards compatibility with the
existing games. This project does not attempt to rebalance or otherwise
modify core gameplay.

Our aims are to:
* Improve the stability of the engine by fixing bugs and improving
  performance.
* Provide a clean base from which new JO and JA code modifications can
  be made.