The magic in testsuite/mk/test.mk for detecting which library ways are available fails badly if the ghc-prim package contains more than one entry in the library-dirs: field, which is currently the case for our ghc port.
Only check for the existance of vertain files in the first entry
of the library-dirs: field to re-enable tons of previously disabled
tests and to get rid of stupid "/bin/sh: [: ... unexpected
operator/operand" errors.

This should be fixed for real at both our side (don't include the
same directory twice in the library-dirs: field) and upstream (don't
rely on the library-dirs: field only containing one directory).