Import fs2open.
The FreeSpace 2 Source Code Project is the project of a group of
programmers currently maintaining and enhancing the game engine for the
space combat simulator FreeSpace 2, developed by Volition. The source
code was released in 2002, and is used by several projects. Most
prominent among these are games based on the Babylon 5 and 2004
Battlestar Galactica universes.

From Thomas Frohwein, thanks!

Works fine for me. ok sthen@